+++ Form/Backend/Type/AuthorType.php
@@ -34,7 +34,7 @@ class AuthorType extends AbstractType
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
         $resolver->setDefaults(array(
-            'data_class'      => 'Desarrolla2\Bundle\BlogBundle\Form\Backend\Model\AuthorModel',
+            'data_class'      => 'Desarrolla2\Bundle\BlogBundle\Entity\Author',
             'csrf_protection' => true,
         ));
     }
