+++ Form/Backend/Type/LinkType.php
@@ -66,7 +66,7 @@ class LinkType extends AbstractType
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
         $resolver->setDefaults(array(
-            'data_class' => 'Desarrolla2\Bundle\BlogBundle\Form\Backend\Model\LinkModel',
+            'data_class' => 'Desarrolla2\Bundle\BlogBundle\Entity\Link',
             'csrf_protection' => true
         ));
     }
